Handle new lines when a key body is passed through TRITON_KEY_MATERIAL.
This commit fixes a problem where the new lines have been passed as-is to the underlying
Nomad job spec template, causing multiple spurious new lines to appear in the middle
resulting in an invalid HCL causing the parser to fail with an error.

An example would be:

  $ export TRITON_KEY_MATERIAL=$(cat /Users/kwilczynski/.ssh/id_rsa)

The above would store the key in a PEM format as-is, with all the new lines, and
it would be enough to cause the problems when the template is being rendered to
use the `--key-material` command line switch with the tsg binary.

Signed-off-by: Krzysztof Wilczynski <kwilczynski@joyent.com>