config: Add configuration of TritonURL and Datacenter names
This commit adjusts the existing configuration of a Triton DC name as well as
introducing the Triton URL. Triton DC name will be used for scoping a TSG
Management key to each datacenter in which Triton operates. The Triton URL is a
URL CloudAPI for processing manual scaling actions. Both are configurable from
within the config file or through env vars `TSG_TRITON_DC` and `TSG_TRITON_URL`
respectively.