config: Introduce auth.Config to configure the authentication process
In order to utilize the authentication process across different environments, we
want to be able to configure some values. Things like Datacenter, TritonURL,
AuthURL, KeyPrefix, and Whitelisting are all values that can be
configured. Managing these became cumbersome and it was easier to add a
configuration struct to handle these. This commit touches a few files but makes
it easier to configure authentication for these values.