Per-DC Keys and Auth configuration (#53)
* auth: Move dc and triton URL setter into NewSession

* auth: Do not return errors through HTTP interface but log instead

This commit changes how authentication errors are returned. Instead of returning
any backend Go error through the HTTP response, we now log them as Debug
logging. This is to potentially lower the exposure of sensitive information
through our authentication process while still giving us the option to debug.

* auth: Add per-DC key creation by naming keys in Triton

* config: Configure the URL used to authenticate against CloudAPI

This commit introduces the ability to configure the URL used to authenticate
against CloudAPI.

There are many types of Triton installations, both public and private. To hedge
against low latency, some large scale installations will want to target their
authentication against the host datacenter of a UFDS master. For everyone else
who does not configure this setting the same configured TritonURL will be used
to authenticate accounts against CloudAPI.

* config: Introduce auth.Config to configure the authentication process

In order to utilize the authentication process across different environments, we
want to be able to configure some values. Things like Datacenter, TritonURL,
AuthURL, KeyPrefix, and Whitelisting are all values that can be
configured. Managing these became cumbersome and it was easier to add a
configuration struct to handle these. This commit touches a few files but makes
it easier to configure authentication for these values.