orchestrator/job_name: change Job Name format (#75)
Previously, we would make an orchestrastion job use the format of
groupname_templateID. Therefore, if the same group was used with a
different templateID, then we would effectively have 2 jobs doing a
similar thing

We would have had to somehow detect the *old* TemplateID and then drain
that job. Which wasn't working

Therefore, as we can only have 1 use of a group name per triton UUID, we
will now formulate the job name as groupname_tritonUUID.

This will mean that for the lifecycle of a job, we will have the same name
format. So even when a templateID changes, there will only be 1 instance
of a job and there is no chance of a duplicate job being created

Nomad will only allow 1 instance of a job running - so we will get an
error if this happens