Fix. Handle creation of duplicate templates and groups without unique constraint.
This commit aims to change the way how we handle creation of duplicate templates
and groups. Previously, we have used an unique constraint over fields like e.g.,
"name", "account_id" and "archived" (for a group), but because constraint was
trying to ensure that all the designated values, upon removing a group and
adding it back with the same name and account ID, the constraint would work
fine, but a deletion of such newly added group would fail, as there was already
a group like that deleted before (or archived, rather), thus the constraint
would trigger an error.

The change here is to add manual existence check during creation in order to
detect a possible none unique template or group being added, and if so, then
an error is returned back.

Additionally, we add number of indexes to all the tables.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>