Add validation to ensure non-empty group name.
This commit adds a group name validation, which is needed when sending a PUT
request against a group as an empty group name would result in a new task being
added on the Nomad side, but without any reference to an actual group name (or
the old group name, rather). This triggers a replacement of existing instances,
since not there is an additional periodic task in Nomad, where all of the new
instances would not have the `tsg.name` tag set. This means, that they can't
be managed any longer.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>