Add validation to ensure the same group name on update.
This commit adds a check to ensure that the name of the group matches the name
already on the record for the given group ID (effectively for the same group),
and issues an error message otherwise.

At the moment without this check, user could submit a PUT request to update an
existing group, but with a different group name. This new name would be completely
ignored by the update to the database, as the match is done on the group ID (UUID),
but a new Nomad job would be created (rather than updated, as we are going
remove-and-create, so effectively a replace, for the batch jobs in Nomad),
resulting in old instances to be removed and new ones would be spun up with
a new `tsg.name` tag, and thus making them unmanageable going forward.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>